Upload book cover

+++ api/app/models/book.rb
@@ -1,6 +1,9 @@
+require 'base64'
+
 class Book < ApplicationRecord
 
-  require 'base64'
+  has_many :borrowings
+  has_many :reviews
 
   BOOKS_ROOT = File.join('public', 'uploads', 'books')
 
@@ -10,15 +13,15 @@ class Book < ApplicationRecord
       'image/png' => 'png'
   }
   IMG2MIME = {
-      :gif => 'image/gif',
-      :jpeg => 'image/jpeg',
-      :png => 'image/png'
+      'gif' => 'image/gif',
+      'jpeg' => 'image/jpeg',
+      'png' => 'image/png'
   }
 
-  attr_writer :cover, :content
-  attr_readonly :cover_path
 
   before_create :set_default_available_count
+  before_create :set_cover_path
+  before_update :set_cover_path
   after_create :save_content
   after_create :save_cover
   after_update :save_content
@@ -26,18 +29,26 @@ class Book < ApplicationRecord
   after_destroy :delete_content
   after_destroy :delete_cover
 
-  # mount_uploader :cover, BookCoverUploader
-  # mount_uploader :file, BookFileUploader
-  has_many :borrowings
-  has_many :reviews
-
   def available?
     self.available_count > 0
   end
 
+  def cover=(cover)
+    @cover = cover
+  end
+
+  def content=(content)
+    @content = content
+  end
+
   def cover
-    img_format = self.cover_path.split('.').last
-    f = File.open(File.join(BOOKS_ROOT, self.cover_path), 'r')
+    p cover_path
+    if cover_path.nil?
+      return nil
+    end
+    img_format = cover_path.split('.').last
+    p img_format
+    f = File.open(self.cover_path, 'r')
     data = Base64.encode64(f.read)
     content = "data:#{IMG2MIME[img_format]};base64,#{data}="
     f.close
@@ -45,6 +56,9 @@ class Book < ApplicationRecord
   end
 
   def content
+    unless File.exists? content_name
+      return nil
+    end
     f = File.open(content_name, 'r')
     base64_content = Base64.encode64(f.read)
     content = "data:application/pdf;base64,#{base64_content}="
@@ -56,6 +70,26 @@ class Book < ApplicationRecord
     File.join(BOOKS_ROOT, "#{self.id}_content.pdf")
   end
 
+  def delete_cover
+    File.delete(File.join(BOOKS_ROOT, self.cover_path))
+  end
+
+  def delete_content
+    File.delete(content_name)
+  end
+
+  def as_json(options = {})
+    super((options || {}).merge({except: [:cover_path]})).merge({cover: cover}).merge({content: content})
+  end
+
+  private
+  def set_cover_path
+    if @cover =~ /data:(\w+\/\w+);base64,([a-zA-Z0-9+\/]+)=/i
+      p 'Set cover path'
+      self.cover_path = cover_name(MIME2IMG[$1])
+    end
+  end
+
   private
   def set_default_available_count
     self.available_count = self.max_count
@@ -63,12 +97,12 @@ class Book < ApplicationRecord
 
   private
   def save_content
-    if content == nil
+    if @content == nil
       return
     end
-    if content =~ /data:(application\/pdf);base64,([a-zA-Z0-9+\/]+)=/i
+    if @content =~ /data:(application\/pdf);base64,([a-zA-Z0-9+\/]+)=/i
       data = $2
-      open(content_name, 'w+') {|f|
+      open(content_name, 'wb+') {|f|
         f << Base64.decode64(data)
       }
     else
@@ -78,13 +112,13 @@ class Book < ApplicationRecord
 
   private
   def save_cover
-    if content == nil
+    if @cover == nil
       return
     end
-    if cover =~ /data:(\w+\/\w+);base64,([a-zA-Z0-9+\/]+)=/i
+    if @cover =~ /data:(\w+\/\w+);base64,([a-zA-Z0-9+\/]+)=/i
       mime_type = $1
       data = $2
-      open(cover_name(MIME2IMG[mime_type]), 'w+') {|f|
+      open(cover_name(MIME2IMG[mime_type]), 'wb+') {|f|
         f << Base64.decode64(data)
       }
     else
@@ -95,15 +129,5 @@ class Book < ApplicationRecord
   private
   def cover_name(img_format)
     File.join(BOOKS_ROOT, "#{self.id}_cover.#{img_format}")
-  end
-
-  private
-  def delete_cover
-    File.delete(File.join(BOOKS_ROOT, self.cover_path))
-  end
-
-  private
-  def delete_content
-    File.delete(content_name)
   end
 end
